103. roguelike step 2 of 3: travel and the mission loop
World map -> travel -> mission -> back to ship -> world map. Travel is
resolved, not flown: every connected player accepts a proposed site (this
system or one hyperlane hop away, the jammer rule kept), and the world
clock goes on by the jump charge plus physics::travel_days of the leg in
one go, the crisis read at the new day before the ship docks or lands.
Only travel moves the world clock; a mission clock (steps since arrival)
times the droid waves, a town's first wave and the class cooldowns. A
mission starts with everyone healed, charges full and dead players bought
back (BUYBACK_COST, longest dead first, no gear, progress kept); machines
pay the Republic's bounty by tier, pending until the site is cleared.
Back to ship recalls the bots; the departure check waits for standing
players aboard and asks everyone before leaving anyone behind. An
uncleared site is restored to its state at mission start, a town left
under attack falls, dead bots cost BOT_DEATH_PENALTY, and the run is lost
when every player's Bim is dead. The old per-step clock is the switch
World::set_free_clock, off in a run and on in the legacy tests.

World::run (run.rs, mission.rs), refusals 83-91, events 93-106,
REFERENCE_CHECKSUM moved, SAVE_VERSION 34, wire::PROTOCOL 26. App: the
world map panel, Back to ship, the departure window, BIMS_MAP/BIMS_DEPART,
duo_resync.sh travel. The spec's commander call-in reinforcements and
"liberation" do not exist in the code and were left out.

+++ crates/world/src/droid.rs
@@ -25,10 +25,11 @@
 //! - The **wave size** is worked out as each wave appears, so a crew
 //!   that has grown richer between waves meets more machines.
 //! - When the last machine of a wave is destroyed and waves are left,
-//!   the next arrives [`data::DROID_REINFORCE_MINUTES`] of world clock
-//!   later. **Never while one is still standing.**
-//! - A wave whose time came while the crew were away is aboard when they
-//!   next dock. Leaving and coming back resets nothing.
+//!   the next arrives [`data::DROID_REINFORCE_STEPS`] of the **mission
+//!   clock** later (feature 103). **Never while one is still standing.**
+//! - Leaving a station before its last wave is destroyed puts it back as
+//!   the crew met it (`crate::run`): the next visit is a fresh fight, its
+//!   count worked out again at the day it is fought on.
 //!
 //! # How many
 //!
@@ -55,9 +56,9 @@ pub struct Infestation {
     /// Which wave is aboard, counting from one; nought before the first
     /// has ever been laid.
     pub wave: u32,
-    /// The clock minute the next wave arrives, or `None` while one is
-    /// still standing or there are none left.
-    pub next_wave: Option<f64>,
+    /// The step of the **mission clock** the next wave arrives at (feature
+    /// 103), or `None` while one is still standing or there are none left.
+    pub next_wave: Option<u64>,
     /// Whether the count has been worked out yet: the crew's first dock
     /// settles it, and nothing settles it twice.
     pub settled: bool,
